athena-platform/ml/requirements.txt
# ===========================================
# ATHENA ML SERVICE - DEPENDENCIES
# ===========================================

# Core ML/Data Science
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
joblib>=1.3.0,<2.0.0

# Deep Learning (optional, for advanced models)
torch>=2.0.0,<3.0.0

# API Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# Configuration
pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Metrics & Monitoring
prometheus-client>=0.19.0,<1.0.0
structlog>=23.2.0,<24.0.0

# HTTP Client (for Node.js callbacks)
httpx>=0.25.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Feature Engineering
category_encoders>=2.6.0,<3.0.0

# Model Persistence
cloudpickle>=3.0.0,<4.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0

# Type Checking
mypy>=1.7.0,<2.0.0

# Development
ipython>=8.18.0,<9.0.0
black>=23.11.0,<24.0.0
ruff>=0.1.6,<1.0.0
